was-develop/requirements.txt
altgraph==0.17
attrs==19.3.0
Automat==0.8.0
bcrypt==3.2.0
blinker==1.4
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
click==8.0.0
colorama==0.4.3
configobj==5.0.6
constantly==15.1.0
contextlib2==21.6.0
cryptography==35.0.0
cvss==2.3
dbus-python==1.2.16
distro==1.4.0
entrypoints==0.3
execnet==1.9.0
Flask==2.0.0
Flask-Cors==3.0.10
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
iniconfig==1.1.1
itsdangerous==2.0.0
Jinja2==3.0.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jumpssh==1.6.5
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==2.0.0
memory-profiler==0.58.0
mock==4.0.3
more-itertools==4.2.0
netifaces==0.10.4
ntlm-auth==1.5.0
oauthlib==3.1.0
packaging==21.0
paramiko==2.7.2
passlib==1.7.4
path==16.0.0
path.py==12.5.0
pbr==5.6.0
pluggy==0.13.1
psutil==5.8.0
py==1.10.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.20.1
pycparser==2.20
PyGObject==3.36.0
PyHamcrest==1.9.0
pyinstaller==4.4
pyinstaller-hooks-contrib==2021.2
PyJWT==1.7.1
pymacaroons==0.13.0
pymongo==3.12.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
pypsexec==0.2.0
pyrsistent==0.15.5
pyserial==3.4
pyspnego==0.1.6
pytest==6.2.4
pytest-shutil==1.7.0
python-decouple==3.5
python-dotenv==0.19.2
python-owasp-zap-v2.4==0.0.18
pywinrm==0.4.2
PyYAML==5.3.1
redis==3.0.1
redis-log-handler==0.0.1.dev32
requests==2.22.0
requests-ntlm==1.1.0
requests-unixsocket==0.2.0
scp==0.13.3
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
smbprotocol==1.5.1
ssh-import-id==5.10
termcolor==1.1.0
testresources==2.0.1
toml==0.10.2
Twisted==18.9.0
urllib3==1.25.8
uWSGI==2.0.20
wadllib==1.3.3
Werkzeug==2.0.0
xmltodict==0.12.0
zipp==1.0.0
zope.interface==4.7.1
python-dateutil==2.8.2
